fixed logging initilization so dogen is installable by pip again easily
